There was an MdB missing in the data of abgeordnetenwatch so I decided to test all the zipcodes of Germany. This revealed that there are zipcodes missing in the dataset and I donot know why. 117 missing zipcodes revealed.
Furthermore this revealed that my code was not casting
a boolean return code in case of an error into a string
so the catched error itself raised an error during
logging. This is now fixed and the missing zipcodes
now return a simple not found.

The missing postcodes are saved into a file to further
analyse them when the time comes.